avoid compile warning

+++ radio/src/gui/colorlcd/model_select.cpp
@@ -271,7 +271,7 @@ class ModelCategoryPageBody : public FormWindow
                       .c_str(),
                   [=] {
                     modelslist.removeModel(category, model);
-                    update(index < category->size() - 1 ? index : index - 1);
+                    update(index < (int)category->size() - 1 ? index : index - 1);
                   });
             });
           }
